requirements.txt
opencv-python
mtcnn
DeepFace
matplotlib
numpy
pillow
tensorflow
keras
streamlit
googlemaps
python-dotenv
scikit-learn
